requirements.txt
pytorch-lightning
datasets
transformers
scikit-learn
wandb
torchmetrics
matplotlib
seaborn
omegaconf
hydra-core
hydra-colorlog
dvc
onnx
onnxruntime
asyncio
fastapi
uvicorn
opentelemetry-distro
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
pycodestyle==2.11.1
isort==5.13.2